+++ tests/examples.py
@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 
 # domain object:
 class Cat:
-    def __init__(self, name):
+    def __init__(self, name: str):
         self.name = name